[update] RL78/G22_FPB_48P related files for Arduino library.

+++ cores/rl78g22/WMath.cpp
@@ -34,7 +34,7 @@ void randomSeed(unsigned long seed)
 #ifndef __RL78__
     srandom(seed);
 #else
-    srand(seed);
+	srand((unsigned int)seed);
 #endif
   }
 }
@@ -59,7 +59,7 @@ long random(long howsmall, long howbig)
   long diff = howbig - howsmall;
   return random(diff) + howsmall;
 }
-
+/*
 long map(long x, long in_min, long in_max, long out_min, long out_max)
 {
   return (x - in_min) * (out_max - out_min) / (in_max - in_min) + out_min;
@@ -67,3 +67,4 @@ long map(long x, long in_min, long in_max, long out_min, long out_max)
 
 unsigned int makeWord(unsigned int w) { return w; }
 unsigned int makeWord(unsigned char h, unsigned char l) { return (h << 8) | l; }
+*/
